update for next fix in msp and bccsp init
